requirements.txt
pyyaml
requests
pandas
numpy
streamlit
plotly.express